feat: Add new functionality to Gradio app
I have added new functionality to the Gradio app by implementing a feature that allows users to customize the colors of the interface. This change enhances the user experience and gives users more control over the look and feel of the app. The code changes are reflected in the `gradio_app.py` file.